cli/requirements.txt
click>=8.0.0
requests>=2.28.0
PyYAML>=6.0
rich>=13.0.0
